# src/app/authentication/docs/swagger.yaml
definitions:
  models.BFFCreateUserRequest:
    properties:
      confirmPassword:
        example: Secure@123
        minLength: 8
        type: string
      email:
        example: arijit@gmail.com
        type: string
      panCard:
        example: EQZRP1234P
        type: string
      password:
        example: Secure@123
        maxLength: 20
        minLength: 8
        type: string
      phoneNumber:
        example: 7568912340
        maximum: 9999999999
        minimum: 1000000000
        type: integer
      username:
        example: Arijit
        maxLength: 32
        minLength: 5
        type: string
    required:
    - confirmPassword
    - email
    - panCard
    - password
    - phoneNumber
    - username
    type: object
  models.BFFSigninUserResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        type: string
      email:
        example: arijit@gmail.com
        type: string
      username:
        example: Arijit
        type: string
    type: object
  models.BFFSinginUserRequest:
    properties:
      email:
        example: arijit@gmail.com
        type: string
      password:
        example: Ari123Jit@
        maxLength: 20
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.ErrorAPIResponse:
    properties:
      error:
        type: string
      errors:
        $ref: '#/definitions/models.ErrorMessage'
    type: object
  models.ErrorMessage:
    properties:
      errorMessage:
        type: string
      key:
        type: string
    type: object
  models.SuccessAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.BFFSigninUserResponse'
      message:
        example: User logged in successfully
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access and refresh tokens
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BFFSinginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
      summary: Sign in a user
      tags:
      - Authentication
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Handles user registration by validating input and storing user
        details
      parameters:
      - description: User Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BFFCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            type: string
        "400":
          description: Invalid input payload
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
      summary: Create a new user
      tags:
      - User
swagger: "2.0"